fix: depositIntoStrategy tests
